+++ src/lib/index.ts
@@ -4,10 +4,12 @@ import DialogOperation from "./DialogOperation.tsx";
 import AccountingDiary from "./AccountingDiary.tsx";
 import AccountingDiaryWrapper from "./AccountingDiaryWrapper.tsx";
 import LedgerView from "./LedgerView.tsx";
+import FilterDropdown from "./FilterDropdown.tsx";
+import DropZoneOverlay from "./DropZone.tsx";
 import { GlobalProvider } from "./context.tsx";
 import { useAccountingDiary } from "./useAccountingDiary.ts";
 
-export { Header, Footer, DialogOperation, AccountingDiary, AccountingDiaryWrapper, LedgerView, GlobalProvider, useAccountingDiary };
+export { Header, Footer, DialogOperation, AccountingDiary, AccountingDiaryWrapper, LedgerView, FilterDropdown, DropZoneOverlay, GlobalProvider, useAccountingDiary };
 export type { IDataItem, IStyleConfig, ILabels, SortField, SortOrder, ViewMode, AccountingDiaryHandle, UseAccountingDiaryOptions, UseAccountingDiaryReturn } from '../types/common';
 export { defaultLabels } from '../types/common';
 export { exportToCSV, exportToExcel, exportToJSON, importFromCSV, importFromJSON } from './helpers/exportUtils';